backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.8.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
hypercorn==0.15.0
idna==3.6
multidict==6.0.4
openai==1.6.1
priority==2.0.0
pydantic==2.5.3
pydantic_core==2.14.6
quart==0.19.4
quart-cors==0.7.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
wsproto==1.2.0
yarl==1.9.4
PyPDF2==3.0.1
python-docx==0.8.11
Pillow==10.1.0
python-dotenv==1.0.0
asyncpg==0.29.0
tenacity==8.0.1